moved code to component Formjs to clear Appjs
